test data type adjust

+++ toffy/normalize_test.py
@@ -370,8 +370,7 @@ def test_normalize_fov(tmpdir):
     # create image data
     fovs, chans = test_utils.gen_fov_chan_names(num_fovs=1, num_chans=3)
     _, data_xr = test_utils.create_paired_xarray_fovs(
-        tmpdir, fovs, chans, img_shape=(10, 10))
-    data_xr = data_xr.astype('float32')
+        tmpdir, fovs, chans, img_shape=(10, 10), dtype='float32')
 
     # create inputs
     norm_vals = np.random.rand(len(chans))
